some problems with file creation fixed and added things to readme

+++ file_creator.py
@@ -1,7 +1,7 @@
 import os
 from pathlib import Path
 
-def make_files(repository="", directory="", files=[], base_dir=""):
+def make_files(repository="", directory="", files=[], base_dir="", soup=None):
     # if repo and dir are given valid
     if repository == "" or directory == "":
         print("No repo or dir given")
@@ -22,8 +22,47 @@ def make_files(repository="", directory="", files=[], base_dir=""):
 
     # create tasks files
     for file_name in files:
-        open(file_name, 'w').close()
+        if not Path(file_name).exists():
+            open(file_name, 'w').close()
 
-    #create the readme file
+    # #create the readme file
+    # with open('README.md', 'w') as f:
+    #     f.write("### {directory}".format(directory=directory))
+    make_readme(directory, soup)
+
+
+def make_readme(directory="", soup=None):
+    all_resources_elements = soup.find_all('div', {'class': 'panel-body'})
+
+    for div in all_resources_elements:
+        if div.find('h2', string='Requirements'):
+            resources_element = div
+            break
+    try:
+        read_or_watch_element = resources_element.find('h2', string='Resources').find_next_sibling('ul')
+    except:
+        read_or_watch_element = ""
+    try:
+        learning_objectives_element = resources_element.find('h2', string='Learning Objectives').find_next_sibling('ul')
+    except:
+        learning_objectives_element = ""
+
+    # Extract Read or Watch
+    if read_or_watch_element != "":
+        resources = read_or_watch_element.find_all('li')
+
+    if learning_objectives_element != "":
+        learning_objectives = learning_objectives_element.find_all('li')
+
+
+    # Write to README.md
     with open('README.md', 'w') as f:
-        f.write("### {directory}".format(directory=directory))
+        f.write("# {directory}\n".format(directory=directory))
+        if learning_objectives_element != "":
+            f.write("\n## Learning Objectives\n")
+            for objective in learning_objectives:
+                f.write(f"{objective}\n")
+        if read_or_watch_element != "":
+            f.write("\n## Resources\n")
+            for resource in resources:
+                f.write(f"{resource}\n")
